+++ demo/theme-blog/src/.vuepress/theme.ts
@@ -154,31 +154,33 @@ export default hopeTheme(
         components: ["Badge", "VPCard"],
       },
 
+      markdownImage: {
+        figure: true,
+        lazyload: true,
+        size: true,
+      },
+
+      markdownMath: true,
+
+      markdownTab: true,
+
       // All features are enabled for demo, only preserve features you need here
       mdEnhance: {
         align: true,
         attrs: true,
         chart: true,
-        codetabs: true,
         component: true,
         demo: true,
         echarts: true,
-        figure: true,
         flowchart: true,
         gfm: true,
-        imgLazyload: true,
-        imgSize: true,
         include: true,
-        katex: true,
         kotlinPlayground: true,
         mark: true,
         markmap: true,
         mermaid: true,
         playground: {
           presets: ["ts", "vue", "unocss"],
-        },
-        revealJs: {
-          plugins: ["highlight", "math", "search", "notes", "zoom"],
         },
         sandpack: true,
         spoiler: true,
@@ -203,7 +205,6 @@ export default hopeTheme(
         ],
         sub: true,
         sup: true,
-        tabs: true,
         tasklist: true,
         vPre: true,
         vuePlayground: true,
@@ -264,6 +265,10 @@ export default hopeTheme(
             },
           ],
         },
+      },
+
+      revealjs: {
+        plugins: ["highlight", "math", "search", "notes", "zoom"],
       },
 
       seo: