+++ packages/create/template/docs/config/multi/theme.ts
@@ -79,16 +79,29 @@ export default hopeTheme({
       components: ["Badge", "VPCard"],
     },
 
-    // All features are enabled for demo, only preserve features you need here
+    // These features are enabled for demo, only preserve features you need here
+    markdownImage: {
+      figure: true,
+      lazyload: true,
+      size: true,
+    },
+
+    // markdownMath: {
+    //   // install katex before enabling it
+    //   type: "katex",
+    //   // or install mathjax-full before enabling it
+    //   type: "mathjax",
+    // },
+
+    // This features is enabled for demo, only preserve if you need it
+    markdownTab: true,
+
+    // These features are enabled for demo, only preserve features you need here
     mdEnhance: {
       align: true,
       attrs: true,
-      codetabs: true,
       component: true,
       demo: true,
-      figure: true,
-      imgLazyload: true,
-      imgSize: true,
       include: true,
       mark: true,
       plantuml: true,
@@ -108,7 +121,6 @@ export default hopeTheme({
       ],
       sub: true,
       sup: true,
-      tabs: true,
       tasklist: true,
       vPre: true,
 
@@ -126,22 +138,11 @@ export default hopeTheme({
       // gfm requires mathjax-full to provide tex support
       // gfm: true,
 
-      // Install katex before enabling it
-      // katex: true,
-
-      // Install mathjax-full before enabling it
-      // mathjax: true,
-
       // Install mermaid before enabling it
       // mermaid: true,
 
       // playground: {
       //   presets: ["ts", "vue"],
-      // },
-
-      // Install reveal.js before enabling it
-      // revealJs: {
-      //   plugins: ["highlight", "math", "search", "notes", "zoom"],
       // },
 
       // Install @vue/repl before enabling it
@@ -206,6 +207,11 @@ export default hopeTheme({
     //       },
     //     ],
     //   },
+    // },
+
+    // install @vuepress/plugin-revealjs and uncomment these if you need slides
+    // revealjs: {
+    //   plugins: ["highlight", "math", "search", "notes", "zoom"],
     // },
   },
 });