+++ packages/md-enhance/src/node/prepare/prepareConfigFile.ts
@@ -1,4 +1,4 @@
-import { getRealPath, isPlainObject } from "@vuepress/helper";
+import { getRealPath } from "@vuepress/helper";
 import type { App } from "vuepress/core";
 
 import type { MarkdownEnhancePluginOptions } from "../options.js";
@@ -14,7 +14,6 @@ export const prepareConfigFile = async (
 ): Promise<string> => {
   const imports = new Set<string>();
   const enhances = new Set<string>();
-  const setups = new Set<string>();
 
   // TODO: Remove this in v2 stable
   // @ts-expect-error: card does not exist
@@ -36,32 +35,6 @@ export const prepareConfigFile = async (
     enhances.add(`app.component("ChartJS", ChartJS)`);
   }
 
-  if (options.codetabs) {
-    imports.add(
-      `import CodeTabs from "${CLIENT_FOLDER}components/CodeTabs.js";`,
-    );
-    enhances.add(`app.component("CodeTabs", CodeTabs);`);
-
-    // TODO: Remove this in v2 stable
-    if (legacy) {
-      imports.add(
-        `import { hasGlobalComponent } from "${getRealPath(
-          "@vuepress/helper/client",
-          url,
-        )}";`,
-      );
-      imports.add(
-        `import { CodeGroup, CodeGroupItem } from "${CLIENT_FOLDER}compact/index.js";`,
-      );
-      enhances.add(
-        `if(!hasGlobalComponent("CodeGroup", app)) app.component("CodeGroup", CodeGroup);`,
-      );
-      enhances.add(
-        `if(!hasGlobalComponent("CodeGroupItem", app)) app.component("CodeGroupItem", CodeGroupItem);`,
-      );
-    }
-  }
-
   if (options.demo) {
     imports.add(
       `import CodeDemo from "${CLIENT_FOLDER}components/CodeDemo.js";`,
@@ -80,9 +53,6 @@ export const prepareConfigFile = async (
     enhances.add(`injectEChartsConfig(app);`);
   }
 
-  if (options.figure)
-    imports.add(`import "${CLIENT_FOLDER}styles/figure.scss";`);
-
   if (status["flowchart"]) {
     imports.add(
       `import FlowChart from "${CLIENT_FOLDER}components/FlowChart.js";`,
@@ -93,29 +63,6 @@ export const prepareConfigFile = async (
 
   if (status["footnote"])
     imports.add(`import "${CLIENT_FOLDER}styles/footnote.scss";`);
-
-  if (options.hint) {
-    imports.add(
-      `import { useHintContainers } from "${CLIENT_FOLDER}composables/useHintContainers.js";`,
-    );
-    imports.add(`import "${CLIENT_FOLDER}styles/hint/index.scss";`);
-    setups.add("useHintContainers();");
-  }
-
-  if (status["imgMark"])
-    imports.add(`import "${CLIENT_FOLDER}styles/image-mark.scss"`);
-
-  if (status["katex"]) {
-    imports.add(`import "${getRealPath("katex/dist/katex.min.css", url)}";`);
-    imports.add(`import "${CLIENT_FOLDER}styles/katex.scss";`);
-
-    if (isPlainObject(options.katex) && options.katex.copy) {
-      imports.add(
-        `import { useKatexCopy } from "${CLIENT_FOLDER}composables/useKatexCopy.js";`,
-      );
-      setups.add(`useKatexCopy();`);
-    }
-  }
 
   if (status["kotlinPlayground"]) {
     imports.add(
@@ -133,8 +80,6 @@ export const prepareConfigFile = async (
     enhances.add(`app.component("MarkMap", MarkMap);`);
   }
 
-  if (status["mathjax"]) imports.add(`import "./mathjax.css";`);
-
   if (status["mermaid"]) {
     imports.add(`import Mermaid from "${CLIENT_FOLDER}components/Mermaid.js";`);
     imports.add(
@@ -151,28 +96,11 @@ export const prepareConfigFile = async (
     enhances.add(`app.component("Playground", Playground);`);
   }
 
-  if (status["revealJs"]) {
-    imports.add(`import "${getRealPath("reveal.js/dist/reveal.css", url)}";`);
-    imports.add(
-      `import RevealJs from "${CLIENT_FOLDER}components/RevealJs.js";`,
-    );
-    imports.add(
-      `import { injectRevealJsConfig } from "${CLIENT_FOLDER}index.js";`,
-    );
-    enhances.add(`injectRevealJsConfig(app);`);
-    enhances.add(`app.component("RevealJs", RevealJs);`);
-  }
-
-  if (options.tabs) {
-    imports.add(`import Tabs from "${CLIENT_FOLDER}components/Tabs.js";`);
-    enhances.add(`app.component("Tabs", Tabs);`);
-  }
-
   if (status["sandpack"]) {
     imports.add(`import { defineAsyncComponent } from "vue";`);
     imports.add(
       `import { LoadingIcon } from "${getRealPath(
-        "vuepress-shared/client",
+        "@vuepress/helper/client",
         url,
       )}";`,
     );
@@ -212,19 +140,16 @@ app.component(
   return app.writeTemp(
     `md-enhance/config.js`,
     `\
-import { defineClientConfig } from "vuepress/client";
 ${Array.from(imports.values()).join("\n")}
 
-export default defineClientConfig({
+export default {
   enhance: ({ app }) => {
 ${Array.from(enhances.values())
-  .map((item) => `    ${item}`)
+  .map((item) => item.split("\n").map((line) => `    ${line}`))
+  .flat()
   .join("\n")}
   },
-  setup: () => {
-${Array.from(setups.values()).join("\n")}
-  }
-});
+};
 `,
   );
 };