+++ packages/theme/src/client/modules/navbar/components/Navbar.ts
@@ -78,11 +78,9 @@ export default defineComponent({
       Links: NavbarLinks,
       Repo: RepoLink,
       Outlook: OutlookButton,
-      Search: hasGlobalComponent("Docsearch")
-        ? resolveComponent("Docsearch")
-        : hasGlobalComponent("SearchBox")
-          ? resolveComponent("SearchBox")
-          : noopComponent,
+      Search: hasGlobalComponent("SearchBox")
+        ? resolveComponent("SearchBox")
+        : noopComponent,
     };
 
     const getNavbarComponent = (component: string): Component | string =>
@@ -98,6 +96,7 @@ export default defineComponent({
           key: "navbar",
           id: "navbar",
           class: ["vp-navbar", { "auto-hide": autoHide.value }],
+          "vp-navbar": "",
         },
         [
           h("div", { class: "vp-navbar-start" }, [