+++ packages/theme/src/node/plugins/getPlugins.ts
@@ -12,11 +12,16 @@ import { getCopyCodePlugin } from "./copyCode.js";
 import { getCopyrightPlugin } from "./copyright.js";
 import { getFeedPlugin } from "./feed.js";
 import { getLinksCheckPlugin } from "./linksCheck.js";
+import { getMarkdownHintPlugin } from "./markdownHint.js";
+import { getMarkdownImagePlugin } from "./markdownImage.js";
+import { getMarkdownMathPlugin } from "./markdownMath.js";
+import { getMarkdownTabPlugin } from "./markdownTab.js";
 import { getMdEnhancePlugin } from "./mdEnhance.js";
 import { getNoticePlugin } from "./notice.js";
 import { getPhotoSwipePlugin } from "./photoSwipe.js";
 import { getPwaPlugin } from "./pwa.js";
 import { getRedirectPlugin } from "./redirect.js";
+import { getRevealJsPlugin } from "./revealjs.js";
 import { getRtlPlugin } from "./rtl.js";
 import { getSearchPlugin } from "./search.js";
 import { getSEOPlugin } from "./seo.js";
@@ -59,7 +64,7 @@ export const getPlugins = (
     getCopyCodePlugin(pluginsOptions.copyCode),
     getCopyrightPlugin(themeData, pluginsOptions.copyright, options.hostname),
     // Seo should work before feed
-    getSEOPlugin(themeData, pluginsOptions, options.hostname),
+    getSEOPlugin(themeData, pluginsOptions.seo, options.hostname),
     getFeedPlugin(
       themeData,
       pluginsOptions.feed,
@@ -67,6 +72,11 @@ export const getPlugins = (
       options.favicon,
       legacy,
     ),
+
+    getMarkdownHintPlugin(pluginsOptions.markdownHint),
+    getMarkdownImagePlugin(pluginsOptions.markdownImage),
+    getMarkdownMathPlugin(pluginsOptions.markdownMath),
+    getMarkdownTabPlugin(pluginsOptions.markdownTab),
     getMdEnhancePlugin(pluginsOptions.mdEnhance, legacy),
     getNoticePlugin(pluginsOptions.notice),
     getPhotoSwipePlugin(pluginsOptions.photoSwipe),
@@ -75,6 +85,7 @@ export const getPlugins = (
     getSitemapPlugin(pluginsOptions.sitemap, options.hostname),
     getRtlPlugin(themeData),
     getRedirectPlugin(pluginsOptions.redirect),
+    getRevealJsPlugin(pluginsOptions.revealjs),
     getWatermarkPlugin(pluginsOptions.watermark),
   ].filter((item) => item !== null) as PluginConfig;